fixed several things to run of dt isntead of frame count - cam. Also tweaked several values to make things smoother.

+++ src/GameObject.h
@@ -104,6 +104,6 @@ public:
 
 	glm::vec3 direction;
 
-	int frame = 0;
+	float frame = 0;
 
 };
